+++ apps/builder/app/builder/features/settings-panel/variable-popover.tsx
@@ -61,7 +61,7 @@ import {
   EditorDialog,
   EditorDialogButton,
   EditorDialogControl,
-} from "~/builder/shared/base-code-editor";
+} from "~/builder/shared/code-editor-base";
 import { ResourceForm } from "./resource-panel";
 import { generateCurl } from "./curl";
